exam-result-evaluator.c
This program is a simple C application that evaluates a student's marks and determines whether they pass or fail.

If the marks are greater than 33 and less than or equal to 100, the student is declared Pass.

If the marks entered are greater than 100, the program shows an error message: "You enter wrong MARKS".

If the marks are less than or equal to 33, the student is declared Fail.

The purpose of this program is to practice conditional statements (if, else if, else) in C language. It also demonstrates basic input handling using scanf and output with printf.